Eden-Multimodal/requirements.txt
torch
Pillow
regex
tqdm
peft
librosa
nltk
gradio
